optimization of functions in mqtt_encode - replace uint32_t with uint16_t

+++ MQTTClient/Inc/mqtt_send.h
@@ -7,7 +7,7 @@
 void send_utf8_encoded_str(struct tcp_pcb* pcb, uint8_t* msg, uint16_t len);
 void send_u16(struct tcp_pcb* pcb, uint16_t* val);
 void send_u8(struct tcp_pcb* pcb, uint8_t* val);
-void send_fixed_header(struct tcp_pcb* pcb, uint8_t ctrl_field, uint32_t remaining_len);
+void send_fixed_header(struct tcp_pcb* pcb, uint8_t ctrl_field, uint16_t remaining_len);
 void send_buffer(struct tcp_pcb* pcb, uint8_t* buffer, uint16_t len);
 
 #endif
